Fix broken parameters merge
