+++ ocdeployer/__main__.py
@@ -15,7 +15,7 @@ import shutil
 import prompter
 import yaml
 
-from ocdeployer.utils import oc, load_cfg_file, get_routes, switch_to_project
+from ocdeployer.utils import object_merge, oc, load_cfg_file, get_routes, switch_to_project
 from ocdeployer.secrets import SecretImporter
 from ocdeployer.deploy import DeployRunner
 
@@ -95,24 +95,6 @@ def list_sets(template_dir, output=None):
         print(yaml.dump(as_dict, default_flow_style=False))
 
 
-def object_merge(old, new):
-    """
-    Recursively merge two data structures
-
-    Thanks rsnyman :)
-    https://github.com/rochacbruno/dynaconf/commit/458ffa6012f1de62fc4f68077f382ab420b43cfc#diff-c1b434836019ae32dc57d00dd1ae2eb9R15
-    """
-    if isinstance(old, list) and isinstance(new, list):
-        for item in old[::-1]:
-            new.insert(0, item)
-    if isinstance(old, dict) and isinstance(new, dict):
-        for key, value in old.items():
-            if key not in new:
-                new[key] = value
-            else:
-                object_merge(value, new[key])
-
-
 def get_variables_data(variables_files):
     variables_data = load_cfg_file(variables_files[0])
 
@@ -120,8 +102,7 @@ def get_variables_data(variables_files):
     if len(variables_files) > 1:
         for var_file in variables_files[1:]:
             merged_file_data = load_cfg_file(var_file)
-            object_merge(variables_data, merged_file_data)
-            variables_data = merged_file_data
+            object_merge(merged_file_data, variables_data)
 
     # Check if there's any variables we need to prompt for
     for section, data in variables_data.items():
